Add folders for classes and scenes and arrange each js file in their respective folder. Add crystal assets and sound. Add the function when collide, it plays the sound and delete the created physics object.
