+++ src/scenes/UIScene.js
@@ -1,5 +1,5 @@
-import {HeroesMenu, ActionsMenu, EnemiesMenu} from "./menu"
-import Message from "./messages";
+import { HeroesMenu, ActionsMenu, EnemiesMenu } from "../classes/Menu";
+import Message from "../classes/Messages";
 
 class UIScene extends Phaser.Scene {
   constructor() {
@@ -111,62 +111,4 @@ class UIScene extends Phaser.Scene {
   }
 }
 
-class Unit extends Phaser.GameObjects.Sprite {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame);
-    this.type = type;
-    this.maxHp = this.hp = hp;
-    this.damage = damage; // default damage
-    this.living = true;
-    this.menuItem = null;
-  }
-
-  setMenuItem(item) {
-    this.menuItem = item;
-  }
-
-  attack(target) {
-    if (target.living) {
-      target.takeDamage(this.damage);
-      this.scene.events.emit(
-        "Message",
-        this.type +
-          " attacks " +
-          target.type +
-          " for " +
-          this.damage +
-          " damage"
-      );
-    }
-  }
-  takeDamage(damage) {
-    this.hp -= damage;
-    if (this.hp <= 0) {
-      this.hp = 0;
-      this.menuItem.unitKilled();
-      this.living = false;
-      this.visible = false;
-      this.menuItem = null;
-    }
-  }
-}
-
-class Enemy extends Unit {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame, type, hp, damage);
-    this.setScale(2);
-  }
-}
-
-class PlayerCharacter extends Unit {
-  constructor(scene, x, y, texture, frame, type, hp, damage) {
-    super(scene, x, y, texture, frame, type, hp, damage);
-    this.flipX = true;
-
-    this.setScale(1.5);
-  }
-}
-
-
-
-export {UIScene, Unit, Enemy, PlayerCharacter}
+export { UIScene };
